\name{process.interface}
\Rdversion{1.1}
\docType{class}
\alias{process.interface-class}
\alias{process.interface}

\title{Class ``process.interface''}
\description{The \code{process.interface} class provides a reference object
interface for any class that extends the background package.
}
\keyword{classes}
