\name{sequential.process}
\Rdversion{1.1}
\docType{class}
\alias{sequential.process-class}
\alias{sequential.process}

\title{Class ``sequential.process''}
\description{The \code{sequential.process} class implements the 
\code{process.interface} and provides sequential execution.  Note this
is intended for debugging.
}
\keyword{classes}
